+++ .gitignore
@@ -1,0 +1,4 @@
+
+assets/
+php/
+.vscode
