+++ song_rcm.html
@@ -1,0 +1,160 @@
+<!DOCTYPE html>
+<html lang="zxx">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="description" content="">
+    <meta name="author" content="">
+    <link rel="icon" href="assets/img/basic/favicon.ico" type="image/x-icon">
+    <title>Record</title>
+    <!-- CSS -->
+    <link rel="stylesheet" href="assets/css/app.css">
+</head>
+
+<body class="sidebar-mini sidebar-collapse theme-dark  sidebar-expanded-on-hover has-preloader">
+    <div id="rekord-app">
+        <!--Sidebar-->
+        <aside class="main-sidebar fixed offcanvas shadow" data-toggle='offcanvas'>
+            <div class="sidebar">
+                <ul class="sidebar-menu">
+                    <li><a class="ajaxifyPage active" href="song_rcm_list.html">
+                            <i class="icon icon-home-1 s-24"></i> <span>Home</span>
+                        </a>
+                    </li>
+                    <li><a class="ajaxifyPage" href="song_rcm_list.html">
+                            <i class="icon icon-layers-1 s-24"></i> <span>recommend_songs</span>
+                        </a>
+                    <li>
+                    <li><a class="ajaxifyPage" href="#">
+                            <i class="icon icon-windows s-24"></i> <span>Playlist</span>
+                        </a>
+                    <li>
+                    <li><a class="ajaxifyPage" href="#">
+                            <i class="icon icon-video-player-2 s-24"></i> <span>Search</span>
+                        </a>
+                    </li>
+                </ul>
+            </div>
+        </aside>
+        <!--Sidebar End-->
+
+        <!--navbar-->
+        <nav class="navbar-wrapper navbar-bottom-fixed shadow">
+            <div class="navbar navbar-expand player-header justify-content-between  bd-navbar">
+                <div class="d-flex align-items-center">
+                    <a href="#" data-toggle="push-menu" class="paper-nav-toggle pp-nav-toggle  paper-nav-toggle-sidenav ml-2 mr-2">
+                        <i></i>
+                    </a>
+                    <a class="navbar-brand d-none d-lg-block" href="index.html">
+                        <div class="d-flex align-items-center s-14 l-s-2">
+                            <span>RECORD</span>
+                        </div>
+                    </a>
+                </div>
+                <!--Top Menu Start -->
+        <div class="navbar-custom-menu">
+            <ul class="nav navbar-nav">
+
+                <!-- Right Sidebar Toggle Button -->
+                <li class="searchOverlay-wrap">
+                    <a href="#" id="btn-searchOverlay" class="nav-link mr-3 btn--searchOverlay no-ajaxy">
+                        <i class="icon icon-search s-24"></i>
+                    </a>
+
+                </li>
+                <!-- User Account-->
+                <li class="dropup custom-dropdown user user-menu ">
+                    <a href="#" class="nav-link" data-toggle="dropdown">
+                        <figure class="avatar">
+                            <img src="assets/img/demo/u7.png" alt="">
+                        </figure>
+                        <i class="icon-more_vert "></i>
+                    </a>
+                    <div class="dropdown-menu p-4 dropdown-menu-right">
+                        <div class="row box justify-content-between my-4">
+                            <div class="col text-center">
+                                <a class="ajaxifyPage" href="saved.html">
+                                    <i class="icon icon-save s-24"></i> <span>Saved</span>
+                                </a>
+                            </div>
+                            <div class="col text-center">
+                                <a class="ajaxifyPage" href="saved.html">
+                                    <i class="icon icon-heart s-24"></i> <span>Favourites</span>
+                                </a>
+                            </div>
+                            <div class="col text-center">
+                                <a class="ajaxifyPage" href="profile.html">
+                                    <i class="icon-user-4  s-24"></i>
+                                    <div class="pt-1">Profile</div>
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+                </li>
+            </ul>
+        </div>
+            </div>
+
+        </nav>
+        
+        <!--Page Content-->
+        <main id="pageContent" class="page has-sidebar">
+            <div class="container-fluid relative animatedParent animateOnce no-p">
+                <div class="animated fadeInUpShort">
+                    <!--Banner Slider-->
+                    <section>
+                        <div class="text-white">
+                            <div class="lightSlider" data-item="1" data-controls="true" data-slide-margin="0"
+                                data-gallery="false" data-pause="8000" data-pauseonhover="true" data-auto="false"
+                                data-pager="false" data-loop="true">
+                                <div class="xv-slide" data-bg-possition="top"
+                                    style="background-image:url('assets/img/demo/b1.jpg');">
+                                    <div class="wrapper has-bottom-gradient">
+                                        <div class="p-md-5 p-3">
+                                            <div class="row">
+                                                <div class="col-12 col-lg-6 fadeInRight animated">
+                                                    <div class="xv-slider-content clearfix color-white">
+                                                        <h1 class="s-64 mt-5 font-weight-lighter"">Recommend</h1>
+                                                        <p class="s-24 font-weight-lighter">좋아하실 만한 노래를 추천했어요</p>
+                                                        <div class="pt-3">
+                                                            <a class="btn btn-primary btn-lg" href="song_rcm_list.html"">다시 검색하기</a>
+                                                        </div>
+                                                    </div>
+                                                </div>
+                                            </div>
+                                        </div>
+                                    </div>
+                                    <div class="bottom-gradient"></div>
+                                </div>
+                            </div>
+                        </div>
+                        <!--slider Wrap-->
+                    </section>
+                    <!--@Banner Slider-->
+
+                    <div class="wrapper p-md-5 p-3  ">
+                        <div class="wrapper">
+                            <div class="row has-items-overlay">
+                                <!--Song Card Object-->
+                                <div id='songs-box2'></div>
+                            </div>
+                        </div>
+
+                    </div>
+                </div>
+            </div>
+        </main>
+        <!--@Page Content-->
+    </div>
+    <!--@#rekord-app-->
+    <!--/#rekord-app -->
+    <script
+        src="https://maps.googleapis.com/maps/api/js?&key=AIzaSyC3YkZNNySdyR87o83QEHWglHfHD_PZqiw&libraries=places"></script>
+    <script src="assets/js/app.js"></script>
+<script src="assets/js/app.js"></script>
+<script src="./static/script/song_rcm_api.js"></script>
+
+
+
+</body>
+</html>
